+++ jaraco/develop/projects-run.py
@@ -1,14 +1,8 @@
 """
-Mechanized merging of an upstream repo across all projects.
-
-Relies on 'known-merge' tool, which must be configured with
-git thus:
-
-[mergetool "known-merge"]
-cmd = py -m jaraco.develop.merge "$BASE" "$LOCAL" "$REMOTE" "$MERGED"
-trustExitCode = true
+Routine to run a command across all projects.
 """
 
+import argparse
 import subprocess
 
 import autocommand
@@ -17,14 +11,33 @@ from . import filters
 from . import git
 
 
-@autocommand.autocommand(__name__)
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    '--keyword',
+    '-k',
+    dest='selectors',
+    type=filters.Keyword,
+    default=filters.Selectors(),
+    action='append',
+)
+parser.add_argument(
+    '--tag',
+    '-t',
+    dest='selectors',
+    type=filters.Tag,
+    default=filters.Selectors(),
+    action='append',
+)
+parser.add_argument('args', nargs='*')
+
+
+@autocommand.autocommand(__name__, parser=parser)
 def main(
-    keyword: filters.Keyword = None,  # type: ignore
-    tag: filters.Tag = None,  # type: ignore
-    *args,
+    selectors: filters.Selectors,
+    args=None,
 ):
-    for project in filter(tag, filter(keyword, git.projects())):
-        print(project)
+    for project in filter(selectors, git.projects()):
+        print(project, flush=True)
         with git.temp_checkout(project, quiet=True):
-            subprocess.Popen(args)
-        print()
+            subprocess.Popen(args).wait()
+        print(flush=True)